+++ src/include/class-wc-retailcrm-uploader.php
@@ -75,13 +75,13 @@ if (class_exists('WC_Retailcrm_Uploader') === false) {
         /**
          * Uploads archive order in CRM
          *
-         * @param null|int $page Number page uploads.
-         * @param array    $ids  Ids orders upload.
+         * @param int|null $page Number page uploads.
+         * @param array $ids  Ids orders upload.
          *
          * @return void|null
          * @throws Exception Invalid argument exception.
          */
-        public function uploadArchiveOrders($page, $ids = [])
+        public function uploadArchiveOrders(?int $page, array $ids = [])
         {
             WC_Retailcrm_Logger::info(__METHOD__, 'Archive order IDs: ' . implode(', ', $ids));
 
@@ -98,11 +98,13 @@ if (class_exists('WC_Retailcrm_Uploader') === false) {
                 $orderIds = $ids;
             }
 
-            foreach ($orderIds as $orderId) {
-                $errorMessage = $this->orders->orderCreate($orderId);
+            if ($orderIds !== []) {
+                foreach ($orderIds as $orderId) {
+                    $errorMessage = $this->orders->orderCreate($orderId);
 
-                if (is_string($errorMessage)) {
-                    $uploadErrors[$orderId] = $errorMessage;
+                    if (is_string($errorMessage)) {
+                        $uploadErrors[$orderId] = $errorMessage;
+                    }
                 }
             }
 
@@ -125,7 +127,7 @@ if (class_exists('WC_Retailcrm_Uploader') === false) {
 
             $users = $this->getCmsUsers($page);
 
-            if (false === empty($users)) {
+            if ($users !== []) {
                 $dataCustomers = [];
 
                 foreach ($users as $user) {
@@ -234,5 +236,47 @@ if (class_exists('WC_Retailcrm_Uploader') === false) {
                 );
             }
         }
+
+        public function uploadConsole($entity, $page = 0)
+        {
+            $ordersPages = (int) ceil($this->getCountOrders() / 50);
+            $customerPages = (int) ceil($this->getCountUsers() / 50);
+
+            try {
+                switch ($entity) {
+                    case 'orders':
+                        $this->archiveUpload('orders', $page, $ordersPages);
+                        break;
+                    case 'customers':
+                        $this->archiveUpload('customers', $page, $customerPages);
+                        break;
+                    case 'full_upload':
+                        $this->archiveUpload('customers', 0, $customerPages);
+                        $this->archiveUpload('orders', 0, $ordersPages);
+                        break;
+                    default:
+                        echo 'Unknown entity: ' . $entity;
+                }
+            } catch (Exception $exception) {
+                echo $exception->getMessage();
+            }
+        }
+
+        public function archiveUpload($entity, $page, $totalPages)
+        {
+            echo $entity . ' uploading started' . PHP_EOL;
+
+            do {
+                if ($entity === 'orders') {
+                    $this->uploadArchiveOrders($page);
+                } elseif ($entity === 'customers') {
+                    $this->uploadArchiveCustomers($page);
+                }
+
+                echo $page . ' page uploaded' . PHP_EOL;
+
+                $page++;
+            } while ($page <= $totalPages);
+        }
     }
-}//end if
+}