+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2024-12-19 4.8.15
+* Fix uploading archive in CRM using console script
+
 ## 2024-11-07 4.8.14
 * The method for determining the stock quantity has been optimized